--- pkg/p2p/internal/overlay/overlay.proto
@@ -1,12 +1,0 @@
-syntax = "proto3";
-
-package overlay;
-
-message OverlayReq {
-    string Address = 1;
-}
-
-message OverlayResp {
-    string Address = 1;
-}
-
